+++ src/app.js
@@ -13,58 +13,29 @@ const routes = require('./routes/v1');
 const { errorConverter, errorHandler } = require('./middlewares/error');
 const ApiError = require('./utils/ApiError');
 const router = require('./routes/v1/user.route');
-
 const app = express();
 if (config.env !== 'test') {
   app.use(morgan.successHandler);
   app.use(morgan.errorHandler);
 }
-
 app.use(bodyparser.urlencoded({extended: false}))
 app.use(bodyparser.json())
-// set security HTTP headers
 app.use(helmet());
-
-// parse json request body
 app.use(express.json());
-
-// parse urlencoded request body
 app.use(express.urlencoded({ extended: true }));
-
-// sanitize request data
 app.use(xss());
 app.use(mongoSanitize());
-
-// gzip compression
 app.use(compression());
-
-
-// enable cors
 app.use(cors());
 app.options('*', cors());
-
-
-// limit repeated failed requests to auth endpoints
 if (config.env === 'production') {
   app.use('/v1/auth', authLimiter);
 }
-
-// v1 api routes
 app.use('/v1', routes);
-
-// send back a 404 error for any unknown api request
 app.use((req, res, next) => {
   next(new ApiError(httpStatus.NOT_FOUND, 'Not found'));
 });
-
-// convert error to ApiError, if needed
 app.use(errorConverter);
-
-// handle error
 app.use(errorHandler);
-
-
-//user-auth
 app.use('/user',router);
-
 module.exports = app;
