+++ src/controllers/index.js
@@ -1,1 +1,2 @@
 module.exports.userController = require('./user.controller');
+module.exports.paymentController = require('./payment.controller');
