+++ src/services/index.js
@@ -1,1 +1,2 @@
 module.exports.userService = require('./user.service');
+module.exports.paymentService = require('./payment.service');
