+++ src/utils/catchAsync.js
@@ -1,5 +1,5 @@
-const catchAsync = (fn) => (req, res, next) => {
-  Promise.resolve(fn(req, res, next)).catch((err) => next(err));
+const catchAsync = (fn) => (req, res) => {
+  Promise.resolve(fn(req, res)).catch((err) => console.log(err));
 };
 
 module.exports = catchAsync;
